requirements.txt
requests
beautifulsoup4
pandas
python-dotenv
supabase
groq
openai
google-genai
playwright
